==> code/03_staar_2016_2017_tables_plots.Rmd <==
---
title: "03_staar_2016_2017_tables_plots"
author: "Tom"
date: "May 7, 2018"
output: html_document
---

# Goals for project

Calculate percentage of students whose **STAAR reading scores** are:  

* Approaching grade level
    * Overall
    * By school type (Elementary, middle, high school)
    * By gender
    * By race
* Meets grade level
    * Overall
    * By school type (E, M, S)
    * By gender
    * By race
* Create a graph    

### Load Libraries

```{r, echo = F}
# load libraries
library(tidyverse)
library(readxl)
library(here)
library(writexl)
library(ggalt)
library(kableExtra)
library(hrbrthemes)
library(ggpubr)
```

## Read in data

```{r}
# read in clean, formatted data
approach_df <- readRDS(here("output_data/approach_df.rds"))
meeting_df <- readRDS(here("output_data/meeting_df.rds"))
```

## Generate summaries by group

```{r}
approaching_summary <- approach_df %>% 
    filter(group_code != "American Indian", group_code != "Pacific Islander") %>% 
    group_by(gradetype, group_code, year_code) %>% 
    add_tally() %>% 
    group_by(gradetype, group_code, year_code, n) %>% 
    summarize(percent_approaching = mean(rate, na.rm = T),
              raw_sum = sum(denom_students, na.rm = T))

meeting_summary <- meeting_df %>% 
    group_by(gradetype, group_code, year_code) %>% 
    filter(group_code != "American Indian", group_code != "Pacific Islander") %>% 
    add_tally() %>% 
    group_by(gradetype, group_code, year_code, n) %>% 
    summarize(percent_meeting = mean(rate, na.rm = T),
              raw_sum = sum(denom_students, na.rm = T))
```

## Create meeting tables

We will create some changes that are not ideal for raw data but are much more readable for tables. We'll save the output into a master excel sheet and then also generate some local html tables to view the data.

```{r}
meeting_table_group <- meeting_summary %>% 
    filter(group_code %in% c("All Students", "Male", "Female", "Econ Disadv", "Special Ed", "At Risk", "ELL")) %>%
    mutate(percent_meeting = scales::percent(percent_meeting)) %>% 
    rename(`Grade-type` = gradetype,
           Group = group_code,
           Year = year_code,
           `Number of Districts` = n,
        `Meeting Grade Level (%)`= percent_meeting,
           `Total Students taking STAAR` = raw_sum) %>%
    ungroup() 
meeting_table_group %>% 
    kable(., "html", caption = "Table 1. Students meeting STAAR Grade Reading level by Group, Grade-type, and Year") %>% 
    kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)

meeting_table_race <- meeting_summary %>% 
    filter(!group_code %in% c("All Students", "Male", "Female", "Econ Disadv", "Special Ed", "At Risk", "ELL")) %>%
    mutate(percent_meeting = scales::percent(percent_meeting)) %>% 
    rename(`Grade-type` = gradetype, 
           Group = group_code,
           Year = year_code,
           `Number of Districts` = n,
            `Meeting Grade Level (%)`= percent_meeting,
           `Total Students taking STAAR` = raw_sum) %>%
    ungroup() 

meeting_table_race %>% 
    kable(., "html", caption = "Table 1. Students meeting STAAR Grade Reading level by Race, Grade-type, and Year") %>% 
    kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)

```

## Create approaching tables


```{r}
approaching_table_group <- approaching_summary %>% 
    filter(group_code %in% c("All Students", "Male", "Female", "Econ Disadv", "Special Ed", "At Risk", "ELL")) %>%
    mutate(percent_approaching = scales::percent(percent_approaching)) %>% 
    rename(`Grade-type` = gradetype,
           Group = group_code,
           Year = year_code,
           `Number of Districts` = n,
        `Approaching Grade Level (%)`= percent_approaching,
           `Total Students taking STAAR` = raw_sum) %>%
    ungroup() 



approaching_table_race <- approaching_summary %>% 
    filter(!group_code %in% c("All Students", "Male", "Female", "Econ Disadv", "Special Ed", "At Risk", "ELL")) %>%
    mutate(percent_approaching = scales::percent(percent_approaching)) %>% 
    rename(`Grade-type` = gradetype, 
           Group = group_code,
           Year = year_code,
           `Number of Districts` = n,
        `Approaching Grade Level (%)`= percent_approaching,
           `Total Students taking STAAR` = raw_sum) %>%
    ungroup() 
```

```{r}
table_group <- left_join(approaching_table_group, meeting_table_group, by = c("Grade-type", "Group", "Year")) %>% 
    select(`Grade-type`:Year,
           `Meeting Grade Level (%)`,
           `Approaching Grade Level (%)`, 
           `Number of Districts.x`, 
           `Total Students taking STAAR.x`, -ends_with(".y")) %>% 
    rename(`Total Students taking STAAR` = `Total Students taking STAAR.x`, `Number of Districts` = `Number of Districts.x`)

table_group %>% 
    kable(., "html", caption = "Table 1. STAAR Reading level by Group, Grade-type, and Year") %>% 
    kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)

table_race <- left_join(approaching_table_race, meeting_table_race, by = c("Grade-type", "Group", "Year")) %>% 
    select(`Grade-type`:Year, 
           `Meeting Grade Level (%)`,
           `Approaching Grade Level (%)`,
           `Number of Districts.x`, 
           `Total Students taking STAAR.x`, -ends_with(".y")) %>% 
    rename(`Total Students taking STAAR` = `Total Students taking STAAR.x`, `Number of Districts` = `Number of Districts.x`)

table_race %>% 
    kable(., "html", caption = "Table 2. STAAR Reading level by Race, Grade-type, and Year") %>% 
    kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)

table_all <- left_join(table_race, table_group, by = c("Grade-type", "Year"))

table_all %>% 
    kable(., "html", caption = "Table 3. STAAR Reading level by Race, Group, Grade-type, and Year") %>% 
    kable_styling(bootstrap_options = c("striped", "hover"), full_width = F)
```

## Save data to Excel and as RDS file

```{r}
list(group = table_group, race = table_race, all = table_all) %>% 
    saveRDS(here("/output_data/STAAR_16_17_reading_summary_tables.RDS"))
# RDS files
list(approaching = approaching_summary, meeting = meeting_summary) %>% 
    writexl::write_xlsx(here("/output_data/STAAR_16_17_reading_summary_tables.xlsx"))
```

### Making graphs

My goal is to graph Racial Achievement gap by percent approaching and percent meeting. I will be focusing on asian, african american, hispanic, and white students as they make up the majority of the students and we have complete data for them.

```{r}
# trial plots, high level overview
ggplot(approaching_summary, aes(x = fct_reorder(group_code, percent_approaching), y = percent_approaching)) +
    geom_col() +
    coord_flip() +
    facet_grid(year_code~gradetype) +
    scale_y_continuous(labels = scales::percent) +
    labs(x = "", 
         y = "Approaching Grade Level for Reading (%)")

ggplot(meeting_summary, aes(x = fct_reorder(group_code, percent_meeting), y = percent_meeting)) +
    geom_col() +
    coord_flip() +
    facet_grid(year_code~gradetype) +
    scale_y_continuous(labels = scales::percent, breaks = seq(0, 1, 0.2)) +
    labs(x = "", 
         y = "\nMeeting Grade Level for Reading (%)",
         title = "Racial achievement Gap")
    
```

```{r}
# Trial plot
ggplot(filter(approaching_summary, as.character(group_code) %in% c("African American", "White", "Hispanic", "Asian")), 
       aes(x = year_code, y = percent_approaching, color = group_code, group = group_code)) +
    geom_path() +
    geom_point(size = 3) +
    facet_grid(~gradetype) +
    scale_y_continuous(labels = scales::percent, breaks = seq(0, 1, 0.2), limits = c(0, 1)) +
    labs(x = "", 
         y = "\nMeeting Grade Level for Reading (%)",
         title = "Racial Achievement gap")

levels(meeting_df$group_code)

```

```{r}
# create the approaching grade standards plot for racial groups (excluding Native American and Pacific Islanders)
# as there is unavailable data or simply no students
(approaching_gap_plot <- ggplot(filter(approaching_summary, group_code %in% c("African American", "White", "Hispanic", "Asian")), 
                                aes(x = year_code, y = percent_approaching, color = group_code, group = group_code)) +
    geom_path() +
    geom_point(size = 3) +
    facet_grid(~gradetype) +
    scale_y_continuous(labels = scales::percent, breaks = seq(0, 1, 0.2), limits = c(0, 1)) +
    labs(x = "", 
         y = "\nApproaching Grade Level for Reading (%)",
         title = "There is a distinct achievement gap in students approaching reading standards\nbetween White/Asian students compared to Hispanic/African American students",
         caption = "Data Source: Teaching Trust") +
    hrbrthemes::theme_ipsum_rc() +
    scale_color_ipsum() +
    theme(legend.title = element_blank(),
          legend.position = "top",
          strip.text = element_text(hjust = .5, size = 12),
          axis.title.y = element_text(vjust = .5, size = 14, face = "bold")))

(meeting_gap_plot <- ggplot(filter(meeting_summary, group_code %in% c("African American", "White", "Hispanic", "Asian")), 
                            aes(x = year_code, y = percent_meeting, color = group_code, group = group_code)) +
    geom_path() +
    geom_point(size = 3) +
    facet_grid(~gradetype) +
    scale_y_continuous(labels = scales::percent, breaks = seq(0, 1, 0.2), limits = c(0, 1)) +
    labs(x = "", 
         y = "\nMeeting Grade Level for Reading (%)",
         title = "There is a distinct achievement gap in students meeting reading standards\nbetween White/Asian students compared to Hispanic/African American students",
         caption = "Data Source: Teaching Trust") +
    hrbrthemes::theme_ipsum_rc() +
    scale_color_ipsum() +
    theme(legend.title = element_blank(),
          legend.position = "top",
          strip.text = element_text(hjust = .5, size = 12),
          axis.title.y = element_text(vjust = .5, size = 14, face = "bold")))
```

# Save plots at a quality level
```{r}
ggsave(here("plots/meeting_gap.png"), meeting_gap_plot, width = 10, height = 8, units = "in", dpi = 1000)
ggsave(here("plots/approaching_gap.png"), approaching_gap_plot, width = 10, height = 8, units = "in", dpi = 1000)
```
